feat: add workspace system info view
